BE/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic>=2.6,<2.9
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
typing_extensions==4.12.2
email-validator==2.2.0
python-dotenv==1.0.1